# docs/api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Banking Platform API
  description: Aggregated API from all microservices
  contact:
    name: Agent Banking Platform Team
servers:
- description: Gateway
  url: http://localhost:8080
- description: Production
  url: https://api.agentbanking.example.com
- description: Sandbox
  url: https://sandbox.agentbanking.example.com
paths:
...
  /api/v1/transactions/{workflowId}/status:
    get:
      tags:
      - orchestrator-controller (orchestrator-service)
      operationId: getTransactionStatus
      summary: Get transaction workflow status
      description: Poll the status of a previously started transaction workflow.
      security:
        - bearerAuth: []
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
        description: The workflow ID returned from the start transaction response
      responses:
        '200':
          description: Transaction status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - Workflow not found
...
  /api/v1/backoffice/ledger-transactions:
    get:
      tags:
      - ledger-controller (ledger-service)
      operationId: getTransactions
      security:
        - bearerAuth: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/settlement:
    get:
      tags:
      - ledger-controller (ledger-service)
      operationId: getSettlement
      security:
        - bearerAuth: []
      parameters:
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/dashboard:
    get:
      tags:
      - ledger-controller (ledger-service)
      operationId: getDashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/agents/{id}:
    get:
      tags:
      - agent-controller (onboarding-service)
      operationId: getAgent
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - agent-controller (onboarding-service)
      operationId: updateAgent
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - agent-controller (onboarding-service)
      operationId: deactivateAgent
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/backoffice/agents:
    get:
      tags:
      - agent-controller (onboarding-service)
      operationId: listAgents
      security:
        - bearerAuth: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - agent-controller (onboarding-service)
      operationId: createAgent
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentExternalRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/kyc/review-queue:
    get:
      tags:
      - onboarding-controller (onboarding-service)
      operationId: getKycReviewQueue
      security:
        - bearerAuth: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags:
      - auth-controller (auth-iam-service)
      operationId: getMyProfile
      security:
        - bearerAuth: []
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user. For external agents, includes linked agentId.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/token:
    post:
      tags:
      - auth-controller (auth-iam-service)
      operationId: authenticateUser
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/password/change:
    post:
      tags:
      - user-management-controller (auth-iam-service)
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/agents/{agentId}/user-status:
    get:
      tags:
      - user-management-controller (auth-iam-service)
      operationId: getAgentUserStatus
      security:
        - bearerAuth: []
      parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentUserStatusResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/agents/{agentId}/create-user:
    post:
      tags:
      - user-management-controller (auth-iam-service)
      operationId: createAgentUser
      security:
        - bearerAuth: []
      parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                fullName:
                  type: string
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/backoffice/transactions/{workflowId}/maker-propose:
    post:
      tags:
      - resolution-controller (orchestrator-service)
      operationId: makerPropose
      summary: Maker proposes resolution for a transaction
      description: First step in four-eyes approval - maker proposes an action (COMMIT, ROLLBACK, ESCALATE)
      security:
        - bearerAuth: []
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MakerProposalRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolutionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/transactions/{workflowId}/checker-approve:
    post:
      tags:
      - resolution-controller (orchestrator-service)
      operationId: checkerApprove
      summary: Checker approves proposed resolution
      description: Second step in four-eyes approval - checker approves the maker's proposal
      security:
        - bearerAuth: []
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckerActionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolutionResponse'
        '400':
          description: Bad Request - Four-eyes violation or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Self approval not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backoffice/transactions/{workflowId}/checker-reject:
    post:
      tags:
      - resolution-controller (orchestrator-service)
      operationId: checkerReject
      summary: Checker rejects proposed resolution
      description: Second step in four-eyes approval - checker rejects the maker's proposal, returning to maker
      security:
        - bearerAuth: []
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckerActionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolutionResponse'
        '400':
          description: Bad Request - Four-eyes violation or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Self approval not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateAgentExternalRequest:
      required:
      - agentCode
      - businessName
      - tier
      - mykadNumber
      - phoneNumber
      - merchantGpsLat
      - merchantGpsLng
      type: object
      properties:
        agentCode:
          type: string
          description: Unique agent code
          pattern: '^AGT-[0-9]{3}$'
        businessName:
          type: string
          description: Business name
          maxLength: 200
        tier:
          type: string
          enum: [MICRO, STANDARD, PREMIER]
          description: Agent tier level
        mykadNumber:
          type: string
          description: MyKad number (NRIC)
          pattern: '^[0-9]{12}$'
        phoneNumber:
          type: string
          description: Contact number
          pattern: '^60[0-9]{8,9}$'
        merchantGpsLat:
          type: number
          description: Merchant GPS latitude
          minimum: -90
          maximum: 90
        merchantGpsLng:
          type: number
          description: Merchant GPS longitude
          minimum: -180
          maximum: 180
        email:
          type: string
          format: email
        address:
          type: string
          maxLength: 500
    
    # Update Agent Request
    UpdateAgentRequest:
      required:
      - businessName
      - merchantGpsLat
      - merchantGpsLng
      - phoneNumber
      - tier
      type: object
      properties:
        businessName:
          type: string
          description: Business name
          maxLength: 200
        tier:
          type: string
          enum: [MICRO, STANDARD, PREMIUM]
          description: Agent tier level
        merchantGpsLat:
          type: number
          description: Merchant GPS latitude
          minimum: -90
          maximum: 90
        merchantGpsLng:
          type: number
          description: Merchant GPS longitude
          minimum: -180
          maximum: 180
        phoneNumber:
          type: string
          description: Contact number
          pattern: '^60[0-9]{8,9}$'
        email:
          type: string
          format: email
        address:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
    
    # Fee Config Request
...
    TransactionResponse:
      type: object
      properties:
        status:
          type: string
          enum: [SUCCESS, PENDING, FAILED]
        transactionId:
          type: string
          format: uuid
        amount:
          type: string
        currency:
          type: string
        transactionType:
          type: string
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        reference:
          type: string
    
    # Fee Config Response
...
    AgentResponse:
      type: object
      properties:
        agentId:
          type: string
          format: uuid
        agentCode:
          type: string
        businessName:
          type: string
        tier:
          type: string
          enum: [MICRO, STANDARD, PREMIUM]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, PENDING, SUSPENDED]
        merchantGpsLat:
          type: number
        merchantGpsLng:
          type: number
        phoneNumber:
          type: string
        email:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    # Balance Response
...
    DashboardResponse:
      type: object
      properties:
        totalAgents:
          type: integer
        activeAgents:
          type: integer
        totalTransactions:
          type: integer
        totalVolume:
          type: string
        successRate:
          type: number
        dailyStats:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              transactionCount:
                type: integer
              volume:
                type: string
    
    # Transaction List Response
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    
    # Settlement Response
    SettlementResponse:
      type: object
      properties:
        settlementId:
          type: string
          format: uuid
        agentId:
          type: string
        date:
          type: string
          format: date
        totalTransactions:
          type: integer
        totalAmount:
          type: string
        commission:
          type: string
        status:
          type: string
    
    # Error Response
    ErrorResponse:
      required:
      - status
      - error
      type: object
      properties:
        status:
          type: string
          enum: [FAILED]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            action_code:
              type: string
              enum: [DECLINE, RETRY, REVIEW]
            trace_id:
              type: string
            timestamp:
              type: string
              format: date-time

    # ============================================================
    # TRANSACTION ORCHESTRATOR SCHEMAS (NEW - 2026-04-05)
    # ============================================================

    # Transaction Type Enum
...
    TransactionStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED, COMPENSATING, UNKNOWN]
          description: Current workflow status
        workflowId:
          type: string
          description: Workflow identifier
        transactionType:
          type: string
          description: Type of transaction
        amount:
          type: number
          format: double
          description: Transaction amount
        customerFee:
          type: number
          format: double
          description: Fee charged to customer
        referenceNumber:
          type: string
          description: External reference number from switch/biller
        errorCode:
          type: string
          description: Error code if transaction failed
        errorMessage:
          type: string
          description: Human-readable error message
        actionCode:
          type: string
          enum: [DECLINE, RETRY, REVIEW]
          description: Recommended action for failed transactions
        completedAt:
          type: string
          format: date-time
          description: Timestamp when transaction completed

    # Force Resolve Request
...
    MakerProposalRequest:
      required:
      - action
      type: object
      properties:
        action:
          type: string
          enum: [COMMIT, ROLLBACK, ESCALATE]
          description: Proposed action for the transaction
        reasonCode:
          type: string
          description: Reason code for the proposal
        reason:
          type: string
          description: Detailed reason for the proposal
        evidenceUrl:
          type: string
          description: URL to supporting evidence

    # Checker Action Request
    CheckerActionRequest:
      required:
      - reason
      type: object
      properties:
        reason:
          type: string
          description: Reason for approve/reject action

    # Resolution Response
    ResolutionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workflowId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        proposedAction:
          type: string
          enum: [COMMIT, ROLLBACK, ESCALATE]
        reasonCode:
          type: string
        reason:
          type: string
        evidenceUrl:
          type: string
        status:
          type: string
          enum: [PENDING_MAKER, PENDING_CHECKER, APPROVED, REJECTED]
        makerUserId:
          type: string
        makerCreatedAt:
          type: string
          format: date-time
        checkerUserId:
          type: string
        checkerAction:
          type: string
        checkerReason:
          type: string
        checkerCompletedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Token Management
    TokenRequest:
      required:
      - username
      - password
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        grantType:
          type: string
          enum: [password, refresh_token]
          default: password
          minLength: 8
          maxLength: 64
    TokenResponse:
      required:
      - access_token
      - refresh_token
      - expires_in
      - token_type
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int32
          description: Token expiry in seconds
        token_type:
          type: string
          enum: [Bearer]
    MyProfileResponse:
      type: object
      properties:
        userId:
          type: string
          nullable: true
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        userType:
          type: string
          enum: [INTERNAL, EXTERNAL]
        status:
          type: string
        agentId:
          type: string
          nullable: true
          description: The linked agent ID for external users
        mustChangePassword:
          type: boolean
        permissions:
          type: array
          items:
            type: string
...
    ChangePasswordRequest:
      required:
      - currentPassword
      - newPassword
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          maxLength: 64
    ChangePasswordResponse:
      type: object
      properties:
        message:
          type: string

    # Agent User Management
    AgentUserStatusResponse:
      type: object
      properties:
        agentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, CREATED, FAILED]
        userId:
          type: string
          format: uuid
        error:
          type: string
    UserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        status:
          type: string
        userType:
          type: string
          enum: [INTERNAL, EXTERNAL]
        mustChangePassword:
          type: boolean
...